# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["Index"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class AuthPayloadResponseDto {\n} "}}}}},"tags":["Auth"]}},"/api/v1/employees":{"post":{"operationId":"EmployeeController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class EmployeeResponseDto {\n} "}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiErrorResponseFor 500"}}}}},"tags":["Employee"]}},"/api/v1/employees/designations/{id}/employee-details":{"get":{"operationId":"EmployeeController_findEmployeeByPosition","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class EmployeeResponseWithChildDto extends EmployeeResponseDto {\n}  [ ] "}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiErrorResponseFor 500"}}}}},"tags":["Employee"]}},"/api/v1/employees/designations/{id}/employee-details/jwt":{"get":{"operationId":"EmployeeController_findEmployeeByPositionWithAuthorization","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class EmployeeResponseWithChildDto extends EmployeeResponseDto {\n}  [ ] "}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiErrorResponseFor 500"}}}}},"tags":["Employee"],"security":[{"auth":[]}]}},"/api/v1/employee/designations":{"post":{"operationId":"DesignationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DesignationRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class DesignationResponseDto {\n} "}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiErrorResponseFor 500"}}}}},"tags":["Employee Designation"]},"get":{"operationId":"DesignationController_find","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class DesignationResponseDto {\n}  [ ] "}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiErrorResponseFor 500"}}}}},"tags":["Employee Designation"]}}},"info":{"title":"QPRO TASK","description":"QPRO API DOC","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"auth":{"type":"apiKey","in":"header","name":"Authorization","scheme":"bearer","bearerFormat":"Bearer"}},"schemas":{"LoginRequestDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AuthPayloadResponseDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"SwaggerBaseApiResponseFor class AuthPayloadResponseDto {\n} ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"$ref":"#/components/schemas/AuthPayloadResponseDto"}},"required":["status","message","errors","payload","data"]},"EmployeeRequestDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"designation_id":{"type":"number"},"parent_employee_id":{"type":"number"}},"required":["name","email","password","designation_id"]},"EmployeeResponseDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"positionId":{"type":"number"},"positionName":{"type":"string"},"child":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponseDto"}}},"required":["id","name","positionId","positionName","child"]},"SwaggerBaseApiResponseFor class EmployeeResponseDto {\n} ":{"type":"object","properties":{"status":{"type":"number","default":201},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"$ref":"#/components/schemas/EmployeeResponseDto"}},"required":["status","message","errors","payload","data"]},"SwaggerBaseApiErrorResponseFor 500":{"type":"object","properties":{"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"data":{"type":"object","default":null},"status":{"type":"number","default":500}},"required":["message","errors","data","status"]},"EmployeeResponseWithChildDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"positionId":{"type":"number"},"positionName":{"type":"string"},"child":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponseDto"}}},"required":["id","name","positionId","positionName","child"]},"SwaggerBaseApiResponseFor class EmployeeResponseWithChildDto extends EmployeeResponseDto {\n}  [ ] ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponseWithChildDto"}}},"required":["status","message","errors","payload","data"]},"DesignationRequestDto":{"type":"object","properties":{"name":{"type":"string"},"identifier":{"type":"string","default":"cto | senior_software_engineer | software_engineer | junior_software_engineer"}},"required":["name","identifier"]},"DesignationResponseDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"identifier":{"type":"string"}},"required":["id","name","identifier"]},"SwaggerBaseApiResponseFor class DesignationResponseDto {\n} ":{"type":"object","properties":{"status":{"type":"number","default":201},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"$ref":"#/components/schemas/DesignationResponseDto"}},"required":["status","message","errors","payload","data"]},"SwaggerBaseApiResponseFor class DesignationResponseDto {\n}  [ ] ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DesignationResponseDto"}}},"required":["status","message","errors","payload","data"]}}}}